requirements.txt
Flask==3.0.3
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
pytz==2024.1
